+++ picasso/zfit.py
@@ -22,30 +22,13 @@ import numba
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-import matplotlib.gridspec as gridspec
 from tqdm import tqdm
-from scipy.ndimage import (
-    affine_transform as _ndi_affine_transform,
-    center_of_mass as _ndi_center_of_mass,
-    gaussian_filter as _ndi_gaussian_filter,
-    maximum_filter as _ndi_maximum_filter,
-)
-from scipy.optimize import curve_fit, minimize_scalar
-from scipy.signal import fftconvolve
-from scipy.spatial.distance import cdist
+from scipy.optimize import minimize_scalar
 
 from . import io, lib, gausslq, gaussmle, __version__
 
 
 plt.style.use("ggplot")
-
-
-# Default bead-detection / matching parameters for `calibrate_affine`.
-_AFFINE_BEAD_MIN_DISTANCE = 15  # px between two peaks
-_AFFINE_BEAD_THRESHOLD_REL = 0.25  # relative intensity threshold (0-1)
-_AFFINE_BEAD_FIT_RADIUS = 6  # half-width of Gaussian fit patch
-_AFFINE_MATCH_MAX_DIST_PX = 40.0  # max distance between matched pair
-_AFFINE_XCORR_HALF_WIDTH = 18  # half-width of bead crop for xcorr
 
 
 def _nan_index(y: lib.FloatArray1D) -> tuple[lib.BoolArray1D, Callable]:
@@ -310,514 +293,6 @@ def calibrate_z(
         plt.savefig(path + ".png", format="png", dpi=300)
 
     plt.show()
-    return calibration
-
-
-def _movie_to_image(movie) -> np.ndarray:
-    """Collapse a picasso movie to a single 2D float32 image normalised
-    to [0, 1]. Multi-frame movies are averaged; single-frame movies are
-    passed through. Frames are read one-at-a-time so the lazy-loading
-    movie classes in ``picasso.io`` don't have to materialise the full
-    stack at once."""
-    n = len(movie)
-    if n == 0:
-        raise ValueError("Movie has zero frames.")
-    if n == 1:
-        img = np.asarray(movie[0], dtype=np.float32)
-    else:
-        acc = np.zeros(np.asarray(movie[0]).shape, dtype=np.float64)
-        for i in range(n):
-            acc += np.asarray(movie[i], dtype=np.float64)
-        img = (acc / n).astype(np.float32)
-    mn, mx = float(img.min()), float(img.max())
-    return (img - mn) / (mx - mn + 1e-12)
-
-
-def _affine_detect_beads(image: np.ndarray) -> np.ndarray:
-    """Detect bead candidates via Gaussian blur + local-maximum filter.
-    Returns (N, 2) array of [row, col] integer coordinates."""
-    blurred = _ndi_gaussian_filter(image, sigma=1.5)
-    size = 2 * _AFFINE_BEAD_MIN_DISTANCE + 1
-    abs_thresh = (
-        _AFFINE_BEAD_THRESHOLD_REL * (blurred.max() - blurred.min())
-        + blurred.min()
-    )
-    is_peak = (_ndi_maximum_filter(blurred, size=size) == blurred) & (
-        blurred > abs_thresh
-    )
-    border = max(5, _AFFINE_BEAD_MIN_DISTANCE // 2)
-    is_peak[:border] = False
-    is_peak[-border:] = False
-    is_peak[:, :border] = False
-    is_peak[:, -border:] = False
-    return np.column_stack(np.where(is_peak))
-
-
-def _affine_fit_gaussian_2d(patch: np.ndarray):
-    """Fit an elliptical 2D Gaussian to a small image patch. Returns
-    (x0, y0, sx, sy, amp, bg) or None on failure."""
-    ny, nx = patch.shape
-    y0g, x0g = np.unravel_index(np.argmax(patch), patch.shape)
-    bg = np.percentile(patch, 15)
-    amp = patch.max() - bg
-
-    def model(xy, x0, y0, sx, sy, amp, bg):
-        x, y = xy
-        return bg + amp * np.exp(
-            -((x - x0) ** 2 / (2 * sx**2) + (y - y0) ** 2 / (2 * sy**2))
-        )
-
-    xg, yg = np.meshgrid(np.arange(nx), np.arange(ny))
-    try:
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            popt, _ = curve_fit(
-                model,
-                (xg.ravel(), yg.ravel()),
-                patch.ravel(),
-                p0=[x0g, y0g, 2.0, 2.0, amp, bg],
-                bounds=(
-                    [0, 0, 0.3, 0.3, 0, -np.inf],
-                    [nx, ny, nx, ny, np.inf, np.inf],
-                ),
-                maxfev=400,
-            )
-        return popt
-    except Exception:
-        return None
-
-
-def _affine_refine_bead_positions(
-    image: np.ndarray, coarse: np.ndarray
-) -> np.ndarray:
-    """Refine coarse bead positions to sub-pixel accuracy via 2D Gaussian
-    fits. Returns (N, 2) array of refined [row, col] coordinates."""
-    ny, nx = image.shape
-    r = _AFFINE_BEAD_FIT_RADIUS
-    refined = []
-    for ry, rx in coarse:
-        y0 = max(0, ry - r)
-        y1 = min(ny, ry + r + 1)
-        x0 = max(0, rx - r)
-        x1 = min(nx, rx + r + 1)
-        patch = image[y0:y1, x0:x1]
-        popt = _affine_fit_gaussian_2d(patch) if patch.size > 0 else None
-        if popt is not None:
-            refined.append([popt[1] + y0, popt[0] + x0])
-        else:
-            cy, cx = _ndi_center_of_mass(patch) if patch.size > 0 else (ry, rx)
-            refined.append([cy + y0, cx + x0])
-    return np.array(refined)
-
-
-def _affine_match_bead_pairs(
-    coords_ref: np.ndarray, coords_mov: np.ndarray
-) -> tuple[np.ndarray, np.ndarray]:
-    """Match beads via mutual nearest-neighbour with a distance threshold.
-    Returns (pairs_ref, pairs_mov), each (M, 2)."""
-    if len(coords_ref) == 0 or len(coords_mov) == 0:
-        return np.empty((0, 2)), np.empty((0, 2))
-    D = cdist(coords_ref, coords_mov)
-    nn_r2m = np.argmin(D, axis=1)
-    nn_m2r = np.argmin(D, axis=0)
-    pairs_r, pairs_m = [], []
-    for i, j in enumerate(nn_r2m):
-        if D[i, j] < _AFFINE_MATCH_MAX_DIST_PX and nn_m2r[j] == i:
-            pairs_r.append(coords_ref[i])
-            pairs_m.append(coords_mov[j])
-    pairs_ref = np.array(pairs_r) if pairs_r else np.empty((0, 2))
-    pairs_mov = np.array(pairs_m) if pairs_m else np.empty((0, 2))
-    return pairs_ref, pairs_mov
-
-
-def _affine_estimate_2d(src: np.ndarray, dst: np.ndarray) -> np.ndarray:
-    """Estimate a 6-DOF 2D affine transform mapping src -> dst from
-    [row, col] correspondences. Returns a 3x3 homogeneous matrix in the
-    (x=col, y=row) convention."""
-    N = len(src)
-    if N < 3:
-        if N == 0:
-            return np.eye(3)
-        tx = float(np.mean(dst[:, 1] - src[:, 1]))
-        ty = float(np.mean(dst[:, 0] - src[:, 0]))
-        return np.array([[1.0, 0.0, tx], [0.0, 1.0, ty], [0.0, 0.0, 1.0]])
-    src_xy = src[:, ::-1].copy()
-    dst_xy = dst[:, ::-1].copy()
-    A_mat = np.zeros((2 * N, 6))
-    b_vec = np.zeros(2 * N)
-    for i, ((xs, ys), (xd, yd)) in enumerate(zip(src_xy, dst_xy)):
-        A_mat[2 * i, :] = [xs, ys, 1, 0, 0, 0]
-        A_mat[2 * i + 1, :] = [0, 0, 0, xs, ys, 1]
-        b_vec[2 * i] = xd
-        b_vec[2 * i + 1] = yd
-    result, _, _, _ = np.linalg.lstsq(A_mat, b_vec, rcond=None)
-    a, b, tx, c, d, ty = result
-    return np.array([[a, b, tx], [c, d, ty], [0.0, 0.0, 1.0]])
-
-
-def _affine_decompose(M: np.ndarray, pixelsize: float | None = None) -> dict:
-    """Decompose the 2x2 linear part of the affine matrix into
-    rotation, anisotropic scaling, and shear via QR factorisation.
-
-    Translations are always reported in pixels (``tx_px``, ``ty_px``).
-    If ``pixelsize`` is given, the nanometre equivalents (``tx_nm``,
-    ``ty_nm``) are added too.
-    """
-    A = M[:2, :2]
-    Q, R = np.linalg.qr(A)
-    signs = np.sign(np.diag(R))
-    Q = Q * signs
-    R = R * signs[np.newaxis, :]
-    rot_deg = np.degrees(np.arctan2(Q[1, 0], Q[0, 0]))
-    scale_x = R[0, 0]
-    scale_y = R[1, 1]
-    shear_deg = np.degrees(np.arctan2(R[0, 1], R[1, 1]))
-    out = {
-        "scale_x": float(scale_x),
-        "scale_y": float(scale_y),
-        "rotation_deg": float(rot_deg),
-        "shear_deg": float(shear_deg),
-        "tx_px": float(M[0, 2]),
-        "ty_px": float(M[1, 2]),
-    }
-    if pixelsize is not None:
-        out["tx_nm"] = float(M[0, 2] * pixelsize)
-        out["ty_nm"] = float(M[1, 2] * pixelsize)
-    return out
-
-
-def _affine_apply(image: np.ndarray, M: np.ndarray) -> np.ndarray:
-    """Warp ``image`` by the affine transform ``M`` (x = col, y = row)
-    using ``scipy.ndimage.affine_transform``. ``M`` is the forward map
-    mov -> ref; we invert it in [row, col] space to feed scipy's pull
-    convention."""
-    M_rc = np.array(
-        [
-            [M[1, 1], M[1, 0], M[1, 2]],
-            [M[0, 1], M[0, 0], M[0, 2]],
-            [0.0, 0.0, 1.0],
-        ]
-    )
-    M_rc_inv = np.linalg.inv(M_rc)
-    A = M_rc_inv[:2, :2]
-    offset = M_rc_inv[:2, 2]
-    return _ndi_affine_transform(
-        image,
-        A,
-        offset=offset,
-        output_shape=image.shape,
-        order=3,
-        mode="constant",
-        cval=0.0,
-    )
-
-
-def _affine_plot_alignment(
-    img_ref: np.ndarray,
-    img_mov: np.ndarray,
-    img_cor: np.ndarray,
-    pairs_ref: np.ndarray,
-    decomp: dict,
-    n_pairs: int,
-    pixelsize: float | None = None,
-    save_path: str = "",
-    ref_path: str = "",
-    cyl_path: str = "",
-) -> None:
-    """Four-panel QC figure: overlay before/after correction and mean
-    per-bead cross-correlation before/after correction.
-
-    If ``pixelsize`` is None, axes are labelled in pixels; otherwise
-    they are scaled to nm and labelled accordingly.
-    """
-    nm = pixelsize if pixelsize is not None else 1.0
-    unit = "nm" if pixelsize is not None else "px"
-
-    def norm(img):
-        mn, mx = img.min(), img.max()
-        return (img - mn) / (mx - mn + 1e-12)
-
-    ref_n = norm(img_ref)
-    mov_n = norm(img_mov)
-    cor_n = norm(img_cor)
-
-    crop_r = _AFFINE_XCORR_HALF_WIDTH
-
-    def bead_xcorr_mean(frame_a, coords_a, frame_b, coords_b):
-        acc, count = None, 0
-        ny, nx = frame_a.shape
-        for (ry_a, rx_a), (ry_b, rx_b) in zip(
-            coords_a.astype(int), coords_b.astype(int)
-        ):
-            ya0 = max(0, ry_a - crop_r)
-            ya1 = min(ny, ry_a + crop_r)
-            xa0 = max(0, rx_a - crop_r)
-            xa1 = min(nx, rx_a + crop_r)
-            yb0 = max(0, ry_b - crop_r)
-            yb1 = min(ny, ry_b + crop_r)
-            xb0 = max(0, rx_b - crop_r)
-            xb1 = min(nx, rx_b + crop_r)
-            pa = frame_a[ya0:ya1, xa0:xa1]
-            pb = frame_b[yb0:yb1, xb0:xb1]
-            if pa.shape[0] < 4 or pb.shape[0] < 4:
-                continue
-
-            def prep(x):
-                x = x - x.mean()
-                s = x.std()
-                return x / (s + 1e-12)
-
-            cc = fftconvolve(prep(pa), prep(pb[::-1, ::-1]), mode="full")
-            cc -= cc.min()
-            if acc is None:
-                acc = np.zeros_like(cc)
-            if cc.shape == acc.shape:
-                acc += cc
-                count += 1
-        if count == 0 or acc is None:
-            s = 4 * crop_r - 1
-            return np.zeros((s, s)), 0
-        return acc / count, count
-
-    def peak_nm(cc):
-        py, px = np.unravel_index(np.argmax(cc), cc.shape)
-        cy_cc, cx_cc = cc.shape[0] // 2, cc.shape[1] // 2
-        r = 5
-        y0 = max(0, py - r)
-        y1 = min(cc.shape[0], py + r + 1)
-        x0 = max(0, px - r)
-        x1 = min(cc.shape[1], px + r + 1)
-        patch = cc[y0:y1, x0:x1]
-        nyp, nxp = patch.shape
-        yg, xg = np.mgrid[0:nyp, 0:nxp].astype(float)
-
-        def g2d(xy, x0, y0, sx, sy, amp, bg):
-            x, y = xy
-            return bg + amp * np.exp(
-                -((x - x0) ** 2 / (2 * sx**2) + (y - y0) ** 2 / (2 * sy**2))
-            )
-
-        try:
-            popt, _ = curve_fit(
-                g2d,
-                (xg.ravel(), yg.ravel()),
-                patch.ravel(),
-                p0=[
-                    nxp / 2,
-                    nyp / 2,
-                    2.0,
-                    2.0,
-                    patch.max() - patch.min(),
-                    patch.min(),
-                ],
-                maxfev=400,
-            )
-            sub_px = x0 + popt[0] - cx_cc
-            sub_py = y0 + popt[1] - cy_cc
-        except Exception:
-            sub_px = float(px - cx_cc)
-            sub_py = float(py - cy_cc)
-        return sub_py * nm, sub_px * nm
-
-    cc_raw, n_raw = bead_xcorr_mean(ref_n, pairs_ref, mov_n, pairs_ref)
-    cc_cor, n_cor = bead_xcorr_mean(ref_n, pairs_ref, cor_n, pairs_ref)
-
-    dy_raw, dx_raw = peak_nm(cc_raw) if n_raw > 0 else (0.0, 0.0)
-    dy_cor, dx_cor = peak_nm(cc_cor) if n_cor > 0 else (0.0, 0.0)
-    off_raw = np.hypot(dy_raw, dx_raw)
-    off_cor = np.hypot(dy_cor, dx_cor)
-
-    fig = plt.figure(figsize=(11, 11))
-    if pixelsize is not None:
-        trans_str = f"Tx={decomp['tx_nm']:.1f} nm  Ty={decomp['ty_nm']:.1f} nm"
-    else:
-        trans_str = f"Tx={decomp['tx_px']:.3f} px  Ty={decomp['ty_px']:.3f} px"
-    title = (
-        f"Alignment check  |  {n_pairs} bead pairs  |  "
-        f"Scale X={decomp['scale_x']:.5f}  Y={decomp['scale_y']:.5f}  "
-        f"Rot={decomp['rotation_deg']:.4f}°  " + trans_str
-    )
-    if ref_path or cyl_path:
-        title += (
-            f"\nref: {os.path.basename(ref_path)}   "
-            f"cyl: {os.path.basename(cyl_path)}"
-        )
-    fig.suptitle(title, fontsize=10, fontweight="bold")
-    gs = gridspec.GridSpec(2, 2, figure=fig, wspace=0.30, hspace=0.25)
-
-    ext = [0, img_ref.shape[1] * nm, img_ref.shape[0] * nm, 0]
-
-    ax = fig.add_subplot(gs[0])
-    ax.imshow(
-        np.clip(np.stack([mov_n, ref_n, mov_n], axis=-1), 0, 1), extent=ext
-    )
-    ax.set_title(
-        "Overlay  BEFORE\nRef (green) | Cylindrical (magenta)",
-        fontsize=9,
-        fontweight="bold",
-    )
-    ax.axis("off")
-
-    ax = fig.add_subplot(gs[1])
-    ax.imshow(
-        np.clip(np.stack([cor_n, ref_n, cor_n], axis=-1), 0, 1), extent=ext
-    )
-    ax.set_title(
-        "Overlay  AFTER\nRef (green) | Corrected (magenta)",
-        fontsize=9,
-        fontweight="bold",
-    )
-    ax.axis("off")
-
-    lag = (np.arange(cc_raw.shape[1]) - cc_raw.shape[1] // 2) * nm
-    e_cc = [lag[0], lag[-1], lag[-1], lag[0]]
-    vmax = max(cc_raw.max(), cc_cor.max(), 1e-12)
-
-    ax = fig.add_subplot(gs[2])
-    im = ax.imshow(
-        cc_raw,
-        cmap="hot",
-        extent=e_cc,
-        vmin=0,
-        vmax=vmax,
-        aspect="equal",
-        interpolation="bilinear",
-    )
-    ax.grid(False)
-    ax.axhline(0, color="cyan", lw=1.0, ls="--", alpha=0.7)
-    ax.axvline(0, color="cyan", lw=1.0, ls="--", alpha=0.7)
-    plt.colorbar(im, ax=ax, fraction=0.046, pad=0.04)
-    ax.set_title(
-        f"Mean bead cross-corr  BEFORE\n"
-        f"peak ({dx_raw:.1f}, {dy_raw:.1f}) {unit}  "
-        f"|offset| = {off_raw:.1f} {unit}",
-        fontsize=9,
-        fontweight="bold",
-    )
-    ax.set_xlabel(f"Δx ({unit})")
-    ax.set_ylabel(f"Δy ({unit})")
-
-    ax = fig.add_subplot(gs[3])
-    im2 = ax.imshow(
-        cc_cor,
-        cmap="hot",
-        extent=e_cc,
-        vmin=0,
-        vmax=vmax,
-        aspect="equal",
-        interpolation="bilinear",
-    )
-    ax.grid(False)
-    ax.axhline(0, color="cyan", lw=1.0, ls="--", alpha=0.7)
-    ax.axvline(0, color="cyan", lw=1.0, ls="--", alpha=0.7)
-    plt.colorbar(im2, ax=ax, fraction=0.046, pad=0.04)
-    ax.set_title(
-        f"Mean bead cross-corr  AFTER\n"
-        f"peak ({dx_cor:.1f}, {dy_cor:.1f}) {unit}  "
-        f"|offset| = {off_cor:.1f} {unit}",
-        fontsize=9,
-        fontweight="bold",
-    )
-    ax.set_xlabel(f"Δx ({unit})")
-    ax.set_ylabel(f"Δy ({unit})")
-
-    fig.tight_layout()
-    if save_path:
-        fig.savefig(save_path, dpi=150, bbox_inches="tight")
-    plt.show()
-
-
-def calibrate_affine(
-    movie_ref,
-    movie_cyl,
-    calibration: dict,
-    pixelsize: float | None = None,
-    ref_path: str = "",
-    cyl_path: str = "",
-    plot_path: str = "",
-) -> dict:
-    """Fit a 6-DOF affine transform that maps the cylindrical-lens bead
-    image into the reference (no-lens) frame and append it to the given
-    3D astigmatism calibration dict.
-
-    The fit is performed in pixel coordinates on a per-pixel mean of
-    each movie. Bead candidates are found by Gaussian-blur + local-max,
-    refined to sub-pixel accuracy by a 2D Gaussian fit, then matched
-    between the two images by mutual nearest neighbour. The affine
-    matrix is solved by 6-DOF linear least squares and decomposed into
-    rotation / anisotropic scale / shear via QR.  TODO: might correct this
-    paragraph
-
-    Parameters
-    ----------
-    movie_ref, movie_cyl : AbstractPicassoMovie
-        In-focus bead movies without (reference) and with the
-        cylindrical lens. If a movie has multiple frames they are
-        averaged; a single-frame movie is used as-is.
-    calibration : dict
-        Existing 3D calibration; an "Affine transform" entry is appended.
-    pixelsize : float, optional
-        Camera pixel size in nm. If given, decomposition translations
-        and the diagnostic plot are converted from pixels to nm. If
-        None (default), values are reported in pixels. Default is None.
-    ref_path, cyl_path : str, optional
-        Paths to the source images, recorded in the calibration for
-        traceability and shown in the diagnostic plot title. Default
-        is "".
-    plot_path : str, optional
-        If given, the diagnostic figure is saved to this path. The
-        figure is always shown interactively. Default is "".
-
-    Returns
-    -------
-    calibration : dict
-        The input calibration augmented with an "Affine transform" key.
-        Use ``io.save_calibration`` to save the result.
-    """
-    img_ref = _movie_to_image(movie_ref)
-    img_cyl = _movie_to_image(movie_cyl)
-
-    coarse_ref = _affine_detect_beads(img_ref)
-    coarse_cyl = _affine_detect_beads(img_cyl)
-    refined_ref = _affine_refine_bead_positions(img_ref, coarse_ref)
-    refined_cyl = _affine_refine_bead_positions(img_cyl, coarse_cyl)
-    pairs_ref, pairs_cyl = _affine_match_bead_pairs(refined_ref, refined_cyl)
-
-    if len(pairs_ref) < 3:
-        raise ValueError(
-            f"Only {len(pairs_ref)} matched bead pair(s) — need >= 3 to "
-            "fit an affine transform. Check the input images / detection "
-            "parameters."
-        )
-
-    M = _affine_estimate_2d(pairs_cyl, pairs_ref)
-    decomp = _affine_decompose(M, pixelsize)
-
-    img_cor = _affine_apply(img_cyl, M)
-    _affine_plot_alignment(
-        img_ref,
-        img_cyl,
-        img_cor,
-        pairs_ref,
-        decomp,
-        n_pairs=len(pairs_ref),
-        pixelsize=pixelsize,
-        save_path=plot_path,
-        ref_path=ref_path,
-        cyl_path=cyl_path,
-    )
-
-    affine_entry = {
-        "Matrix": [[float(v) for v in row] for row in M],
-        "Direction": "cylindrical -> reference (x = col, y = row)",
-        "Reference image": ref_path or "N/A",
-        "Cylindrical image": cyl_path or "N/A",
-        "Bead pairs": int(len(pairs_ref)),
-        "Decomposition": decomp,
-    }
-    if pixelsize is not None:
-        affine_entry["Pixelsize (nm)"] = float(pixelsize)
-    calibration["Affine transform"] = affine_entry
     return calibration
